hiprtc: vs: copy additional libraries to output directory
Besides hiprtc*.dll, hiprtc-builtins*.dll and amd_comgr*.dll is also required.

In this commit, the PostBuildAction is also replaced by a Content statement. This
provides more robustness in the build system, as using this method there will be
no race conditions if two targets copy the same file to the output directory.