ci: do not rely on PATH
Removing the ROCm binary path from the global PATH allows us to check
that the examples do not depend on this. The CMakeFile for each example
does currently not copy the libraries to the build directory, and so
for running the tests using CTest the library path is still required.