Build improvements (#30)
* hiprtc: vs: copy additional libraries to output directory

Besides hiprtc*.dll, hiprtc-builtins*.dll and amd_comgr*.dll is also required.

In this commit, the PostBuildAction is also replaced by a Content statement. This
provides more robustness in the build system, as using this method there will be
no race conditions if two targets copy the same file to the output directory.

* Resolve "Multi-GPU error occured while hip_opengl_interop sample application is executed"

* vs: copy dynamic libraries for other examples as well

If a library links with a dynamic library, the corresponding dll file
should be placed in the output directory so that the library can be found
even if the ROCm binary path is not added to the system PATH. The solution
used is the same as that used for the HIPRTC example.

* cmake: disable static_host_library on windows

The HIP SDK does at this time not ship with a binary for CMAKE_AR, which is required
to compile this example.

* ci: do not rely on PATH

Removing the ROCm binary path from the global PATH allows us to check
that the examples do not depend on this. The CMakeFile for each example
does currently not copy the libraries to the build directory, and so
for running the tests using CTest the library path is still required.

* Resolve "cmake: windows + nvidia does not pass hip includes to include list"

* Resolve "cmake: make examples use install()"

* Resolve "hipBLAS examples' Makefiles do not work with CUDA"

* cuda dockerfile: replace ARG with GID

* Resolve "Windows CI with Nvidia"

* Fix VS copied dependencies for HIP

---------

Co-authored-by: Nara Prasetya <nara@streamhpc.com>
Co-authored-by: Máté Ferenc Nagy-Egri <mate@streamhpc.com>